+++ CHANGELOG.md
@@ -14,6 +14,10 @@ These changes are available on the `master` branch, but have not yet been releas
 
 - Added `Member.vr_status` property.
   ([#3328](https://github.com/Pycord-Development/pycord/pull/3328))
+- Added `ChannelFlags.is_spoiler_channel` and `.is_spoiler()` function to all applicable
+  channel types. ([#3252](https://github.com/Pycord-Development/pycord/pull/3252))
+- Added `spoiler` parameter to the `edit()` function of all applicable channel types.
+  ([#3252](https://github.com/Pycord-Development/pycord/pull/3252))
 
 ### Changed